separate class definition and implementation
